requirements.txt
# Core web framework
flask==2.2.3
flask-cors==6.0.1
Werkzeug==2.2.3
Jinja2==3.1.2

# Environment variables
python-dotenv==1.0.0

# Data processing and machine learning
scikit-learn==1.2.2
numpy==1.24.2
pandas==2.0.0

# AI integration
openai==0.27.4

# Password strength evaluation
zxcvbn==4.5.0

# Visualization support (optional, for development)
matplotlib==3.7.1

# Security utilities
cryptography==39.0.2

# Web utilities
requests==2.28.2 

